+++ R/networks.R
@@ -8,20 +8,18 @@
 #' riem_networks()
 #' }
 riem_networks <- function(){
-  resp <- httr::GET("http://mesonet.agron.iastate.edu/api/1/networks.json")# nolint
-  httr::stop_for_status(resp)
+  resp <- perform_riem_request(path = "api/1/networks.json")
 
-  content <- jsonlite::fromJSON(
-    httr::content(
-      resp, as ="text"
-      )
-    )
+  httr2::resp_check_status(resp)
 
-  names <- content$data$name
-  codes <- content$data$id
-  whichASOS <- grepl("ASOS", codes) | grepl("AWOS", codes)
-  codes <- codes[whichASOS]
-  names <- names[whichASOS]
-  tibble::tibble(code = codes,
-                 name = names)
+  content <- httr2::resp_body_json(resp)
+
+  networks_data <- tibble::tibble(
+    code = purrr::map_chr(content[["data"]], "id"),
+    name = purrr::map_chr(content[["data"]], "name")
+  )
+
+  is_asos_or_awos <- grepl("A[SW]OS", networks_data$code)
+
+  networks_data[is_asos_or_awos,]
 }