+++ R/stations.R
@@ -1,5 +1,7 @@
 #' Function for getting stations of an ASOS network
 #'
+#' @importFrom rlang %||%
+#'
 #' @param network A single network code, see riem_networks() for finding the code corresponding to a name.
 #' @return a data.frame (tibble tibble) with the id, name, longitude (lon) and latitude (lat) of each station in the network.
 #' @details You can see a map of stations in a network at \url{https://mesonet.agron.iastate.edu/request/download.phtml}.
@@ -11,32 +13,33 @@
 #' }
 riem_stations <- function(network = NULL){
 
-  if(is.null(network)){
-    stop("Please provide a network code",
-         call. = FALSE)
-  }
+  valid_network_code <- !is.null(network) && (network %in% riem_networks()$code)
+  if(!valid_network_code) error_invalid_network(network)
 
-  if(!(network %in% riem_networks()$code)){
-    stop(paste0(network,
-                " is not a valid network code. See riem_networks()"),
-         call. = FALSE) # nolint
-  }
-
-  link <- paste0("http://mesonet.agron.iastate.edu/api/1/network/", network, ".json")
-
-  resp <- httr::GET(link)
-  httr::stop_for_status(resp)
-
-  content <- jsonlite::fromJSON(
-    httr::content(
-      resp, as ="text"
-    )
+  resp <- perform_riem_request(
+    path = sprintf("api/1/network/%s.json", network)
   )
 
-  results <- tibble::as_tibble(content$data)
+  httr2::resp_check_status(resp)
+
+  content <- httr2::resp_body_json(resp)
+
+  results <- content[["data"]] %>%
+    purrr::map_df(function(x) tibble::as_tibble(purrr::compact(x)))
+
+
   results <- results[, !names(results) == "combo"]
   results$lon <- as.numeric(results$longitude)
   results$lat <- as.numeric(results$latitude)
 
   return(results)
+}
+
+error_invalid_network <- function(network) {
+  rlang::abort(
+    c(
+      x = sprintf("%s is an invalid network code.", as.character(network %||% "null")),
+      i = "See riem_networks() for valid codes."
+    )
+  )
 }