+++ tests/testthat/test-networks.R
@@ -1,10 +1,17 @@
-library("riem")
-context("networks")
 test_that("riem_networks returns the right output",{
-  vcr::use_cassette("networks", {
-  output <- riem_networks()
+  httptest2::with_mock_dir(file.path("fixtures", "networks"), {
+    output <- riem_networks()
   })
-  expect_is(output, "tbl_df")
-  expect_is(output$code, "character")
-  expect_is(output$name, "character")
+  expect_s3_class(output, "tbl_df")
+  expect_type(output$code, "character")
+  expect_type(output$name, "character")
 })
+
+test_that("riem_networks errors if API error",{
+  my_mock <- function(req) {
+    httr2::response(status_code = 502)
+  }
+  httr2::local_mock(my_mock)
+  expect_snapshot_error(riem_networks())
+})
+