+++ NEWS.md
@@ -1,5 +1,7 @@
 # riem (development version)
 
+* Switches to httr2 and httptest2 under the hood.
+
 # riem 0.2.0
 
 * Switches to newer IEM metadata web services (#35, @akrherz)